Divided into steps and footer
